+++ src/app/pages/auth/services/auth.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { AuthToken } from '../../models/auth-token.model';
-import Endpoints from '../../../../shared/enums/endpoints';
+import { AuthToken } from '../models/auth-token.model';
+import Endpoints from '../../../shared/enums/endpoints';
 
 @Injectable({
   providedIn: 'root',
